Update README.md
In this project, I conducted exploratory data analysis on the "MEPS HC-228 2021 Full Year Population Characteristics" dataset. The goal was to investigate the correlation between demographic characteristics and diabetes. Specifically, I explored the relationship between age, sex, race, ethnicity, employment status, education, country of birth, and the prevalence of diabetes.

I cleaned and analyzed the dataset using R and employed various packages such as dplyr, tidyverse, ggplot2, psych, sqldf, and utils. I performed data cleaning steps to select relevant variables and filter the dataset to include only individuals with diabetes. I calculated age based on year of birth to further analyze the impact of age on diabetes occurrences.

The project focused on identifying if there was a correlation between race and diabetes, as well as understanding the role of key demographic variables in diabetes instances. I visualized the findings using 4-5 charts and summarized the key insights. Additionally, I explored the variation and covariation within and between variables to gain a comprehensive understanding of the data.